Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Might want to do this on Aria and Legend as well. It avoids the KGSL MMU page faults. I'm pretty sure it's an alignment issue in our gralloc, but HTC seems to have worked around it. It should at the very least reduce logspam, and hopefully put an end to the dreaded scheduling-while-atomic problems during KGSL cache flushes.

Patch-set: 1