Update patch set 2
Patch Set 2:

To clarify my previous comment since it made next to no sence:

shouldEnableMenuKey uses the inverse of this property. If the box is unchecked in CMParts and this returns false, making it true when shouldEnableMenuKey returns its result, then menu will ignore the user's choice and enable the menu key. This patch only works under the condition that the phone already has the screen woken up to LockScreen but will bypass this when using menu to wake the phone, I'm hoping to track down why today and submit a separate patch for that issue.

Patch-set: 2